Prevent deletion of the last annotation for a model.

+++ src/store/filestore.rs
@@ -44,6 +44,17 @@ impl Store for LocalFileStore {
 
     fn delete_annotation<T>(&self, name: &str, version: &str) -> Result<()> {
         let hash = self.lookup_hash::<T>(name, version)?;
+        let (count, _) = Self::parse_annotation_path(
+            &self.make_path::<T>(&hash, Self::make_annotation_relpath("*", "*")),
+        )?
+        .size_hint();
+        if count == 1 {
+            return Err(OrcaError::from(Kind::DeletingLastAnnotation(
+                get_type_name::<T>(),
+                name.to_owned(),
+                version.to_owned(),
+            )));
+        }
         let annotation_file =
             self.make_path::<T>(&hash, &Self::make_annotation_relpath(name, version));
         fs::remove_file(&annotation_file)?;
